+++ src/db/index.ts
@@ -1,8 +1,4 @@
-import { neon } from "@neondatabase/serverless";
-import { config } from "dotenv";
-import { drizzle } from "drizzle-orm/neon-http";
-
-config({ path: ".env" });
-const sql = neon(process.env.DATABASE_URL!);
+import { sql } from "@vercel/postgres";
+import { drizzle } from "drizzle-orm/vercel-postgres";
 
 export const db = drizzle({ client: sql });