+++ src/db/schema.ts
@@ -1,4 +1,17 @@
-import { boolean, pgTable, text, timestamp } from "drizzle-orm/pg-core";
+import {
+  boolean,
+  pgEnum,
+  pgTable,
+  text,
+  timestamp,
+  uuid,
+} from "drizzle-orm/pg-core";
+
+export const rolesEnum = pgEnum("role", [
+  "adminRole",
+  "authorRole",
+  "userRole",
+]);
 
 export const user = pgTable("user", {
   id: text("id").primaryKey(),
@@ -14,7 +27,7 @@ export const user = pgTable("user", {
   updatedAt: timestamp("updated_at")
     .$defaultFn(() => new Date())
     .notNull(),
-  role: text("role"),
+  role: rolesEnum("role").default("userRole"),
   banned: boolean("banned"),
   banReason: text("ban_reason"),
   banExpires: timestamp("ban_expires"),
@@ -59,4 +72,21 @@ export const verification = pgTable("verification", {
   expiresAt: timestamp("expires_at").notNull(),
   createdAt: timestamp("created_at").$defaultFn(() => new Date()),
   updatedAt: timestamp("updated_at").$defaultFn(() => new Date()),
+});
+
+export const post = pgTable("post", {
+  id: uuid("id").primaryKey().defaultRandom(),
+  slug: text("slug").unique(),
+  title: text("title"),
+  description: text("description"),
+  content: text("content"),
+  authorId: text("author_id").references(() => user.id),
+  published: boolean("published").notNull().default(false),
+  createdAt: timestamp("created_at")
+    .notNull()
+    .$defaultFn(() => new Date()),
+  updatedAt: timestamp("updated_at")
+    .notNull()
+    .$defaultFn(() => new Date())
+    .$onUpdate(() => new Date()),
 });